+++ CHANGES.rst
@@ -9,6 +9,8 @@
 - Add an ``ansi`` extra to install ``erbsland-sphinx-ansi``. Version
   1.2.4 or later is required.
 - Explicitly list ``docutils`` as a dependency.
+- Add the ability to specify a ``:class:`` argument to set CSS classes
+  on programoutput nodes.
 
 
 0.19 (2026-02-20)